+++ tokens/token-fundraiser/anchor/programs/fundraiser/src/state/fundraiser.rs
@@ -8,6 +8,6 @@ pub struct Fundraiser {
     pub amount_to_raise: u64,
     pub current_amount: u64,
     pub time_started: i64,
-    pub duration: u8,
+    pub duration: u16,
     pub bump: u8,
 }
